chore: improve Popper exports
